+++ Sources/TuistLoader/Models+ManifestMappers/Target+ManifestMapper.swift
@@ -57,11 +57,10 @@ extension TuistGraph.Target {
             generatorPaths: generatorPaths
         )
 
-        var (resources, resourcesPlaygrounds, invalidResourceGlobs) = try resourcesAndPlaygrounds(
+        let (resources, resourcesPlaygrounds, resourcesCoreDatas, invalidResourceGlobs) = try resourcesAndPlaygrounds(
             manifest: manifest,
             generatorPaths: generatorPaths
         )
-        resources = resourcesFlatteningBundles(resources: resources)
 
         if !invalidResourceGlobs.isEmpty {
             throw TargetManifestMapperError.invalidResourcesGlob(targetName: name, invalidGlobs: invalidResourceGlobs)
@@ -79,7 +78,7 @@ extension TuistGraph.Target {
 
         let coreDataModels = try manifest.coreDataModels.map {
             try TuistGraph.CoreDataModel.from(manifest: $0, generatorPaths: generatorPaths)
-        }
+        } + resourcesCoreDatas.map { try TuistGraph.CoreDataModel.from(path: $0) }
 
         let scripts = try manifest.scripts.map {
             try TuistGraph.TargetScript.from(manifest: $0, generatorPaths: generatorPaths)
@@ -124,7 +123,12 @@ extension TuistGraph.Target {
         manifest: ProjectDescription.Target,
         generatorPaths: GeneratorPaths
         // swiftlint:disable:next large_tuple
-    ) throws -> (resources: [TuistGraph.ResourceFileElement], playgrounds: [AbsolutePath], invalidResourceGlobs: [InvalidGlob]) {
+    ) throws -> (
+        resources: [TuistGraph.ResourceFileElement],
+        playgrounds: [AbsolutePath],
+        coreDatas: [AbsolutePath],
+        invalidResourceGlobs: [InvalidGlob]
+    ) {
         let resourceFilter = { (path: AbsolutePath) -> Bool in
             TuistGraph.Target.isResource(path: path)
         }
@@ -132,6 +136,7 @@ extension TuistGraph.Target {
         var invalidResourceGlobs: [InvalidGlob] = []
         var resourcesWithoutPlaygrounds: [TuistGraph.ResourceFileElement] = []
         var playgrounds: Set<AbsolutePath> = []
+        var coreDatas: Set<AbsolutePath> = []
 
         let allResources = try (manifest.resources?.resources ?? []).flatMap { manifest -> [TuistGraph.ResourceFileElement] in
             do {
@@ -146,40 +151,28 @@ extension TuistGraph.Target {
             }
         }
 
-        allResources.forEach { fileElement in
-            switch fileElement {
-            case .folderReference: resourcesWithoutPlaygrounds.append(fileElement)
-            case let .file(path, _):
-                if path.pathString.contains(".playground/") {
-                    playgrounds.insert(path.upToComponentMatching(extension: "playground"))
-                } else {
-                    resourcesWithoutPlaygrounds.append(fileElement)
+        allResources
+            .forEach { fileElement in
+                switch fileElement {
+                case .folderReference: resourcesWithoutPlaygrounds.append(fileElement)
+                case let .file(path, _):
+                    if path.extension == "playground" {
+                        playgrounds.insert(path)
+                    } else if path.extension == "xcdatamodeld" {
+                        coreDatas.insert(path)
+                        resourcesWithoutPlaygrounds.append(fileElement)
+                    } else {
+                        resourcesWithoutPlaygrounds.append(fileElement)
+                    }
                 }
             }
-        }
 
         return (
             resources: resourcesWithoutPlaygrounds,
             playgrounds: Array(playgrounds),
+            coreDatas: Array(coreDatas),
             invalidResourceGlobs: invalidResourceGlobs
         )
-    }
-
-    fileprivate static func resourcesFlatteningBundles(resources: [TuistGraph.ResourceFileElement])
-        -> [TuistGraph.ResourceFileElement]
-    {
-        Array(resources.reduce(into: Set<TuistGraph.ResourceFileElement>()) { flattenedResources, resourceElement in
-            switch resourceElement {
-            case let .file(path, _):
-                if path.pathString.contains(".bundle/") {
-                    flattenedResources.formUnion([.file(path: path.upToComponentMatching(extension: "bundle"))])
-                } else {
-                    flattenedResources.formUnion([resourceElement])
-                }
-            case .folderReference:
-                flattenedResources.formUnion([resourceElement])
-            }
-        })
     }
 
     fileprivate static func sourcesAndPlaygrounds(
@@ -204,8 +197,8 @@ extension TuistGraph.Target {
         } ?? [])
 
         allSources.forEach { sourceFile in
-            if sourceFile.path.pathString.contains(".playground/") {
-                playgrounds.insert(sourceFile.path.upToComponentMatching(extension: "playground"))
+            if sourceFile.path.extension == "playground" {
+                playgrounds.insert(sourceFile.path)
             } else {
                 sourcesWithoutPlaygrounds.append(sourceFile)
             }