+++ Sources/TuistSupport/Extensions/AbsolutePath+Extras.swift
@@ -99,24 +99,6 @@ extension AbsolutePath {
         return ancestorPath
     }
 
-    public func upToComponentMatching(regex: String) -> AbsolutePath {
-        if isRoot { return self }
-        if basename.range(of: regex, options: .regularExpression) == nil {
-            return parentDirectory.upToComponentMatching(regex: regex)
-        } else {
-            return self
-        }
-    }
-
-    public func upToComponentMatching(extension: String) -> AbsolutePath {
-        if isRoot { return self }
-        if self.extension == `extension` {
-            return self
-        } else {
-            return parentDirectory.upToComponentMatching(extension: `extension`)
-        }
-    }
-
     public var upToLastNonGlob: AbsolutePath {
         guard let index = components.firstIndex(where: { $0.isGlobComponent }) else {
             return self