+++ Tests/TuistSupportTests/Extensions/AbsolutePath+ExtrasTests.swift
@@ -68,26 +68,4 @@ final class AbsolutePathExtrasTests: TuistUnitTestCase {
             ))
         )
     }
-
-    func test_upToComponentMatchingRegex() throws {
-        // Given
-        let path = AbsolutePath("/path/to/sources/Playground.playground/Content.swift")
-
-        // When
-        let got = path.upToComponentMatching(regex: ".+\\.playground")
-
-        // Then
-        XCTAssertEqual(got, "/path/to/sources/Playground.playground")
-    }
-
-    func test_upToComponentMatchingExtension() throws {
-        // Given
-        let path = AbsolutePath("/path/to/sources/Playground.playground/Content.swift")
-
-        // When
-        let got = path.upToComponentMatching(extension: "playground")
-
-        // Then
-        XCTAssertEqual(got, "/path/to/sources/Playground.playground")
-    }
 }